test: move delete-payments cron test to new path
