+++ test/integration/02-user-wallet/02-send-lightning.spec.ts
@@ -5,7 +5,7 @@ import { getMidPriceRatio } from "@app/shared"
 
 import { delete2fa } from "@app/users"
 
-import { getDisplayCurrencyConfig, getLocale, TWO_MONTHS_IN_MS } from "@config"
+import { getDisplayCurrencyConfig, getLocale } from "@config"
 
 import { FEECAP_PERCENT, toSats } from "@domain/bitcoin"
 import {
@@ -1901,76 +1901,6 @@ describe("Delete payments from Lnd - Lightning Pay", () => {
     // Check that deleting missing payment doesn't return error
     const deletedAttempt = await lndService.deletePaymentByHash({ paymentHash })
     expect(deletedAttempt).not.toBeInstanceOf(Error)
-  })
-
-  it("runs delete-payment cronjob", async () => {
-    // Create payment
-    const { request, secret, id } = await createInvoice({ lnd: lndOutside1 })
-    const paymentHash = id as PaymentHash
-    const revealedPreImage = secret as RevealedPreImage
-
-    const paymentResult = await Payments.payNoAmountInvoiceByWalletId({
-      paymentRequest: request as EncodedPaymentRequest,
-      memo: null,
-      amount: amountInvoice,
-      senderWalletId: walletIdB,
-      senderAccount: accountB,
-    })
-    if (paymentResult instanceof Error) throw paymentResult
-    expect(paymentResult).toBe(PaymentSendStatus.Success)
-
-    const lndService = LndService()
-    if (lndService instanceof Error) return lndService
-
-    // Confirm payment exists in lnd
-    let retrievedPayment = await lndService.lookupPayment({ paymentHash })
-    expect(retrievedPayment).not.toBeInstanceOf(Error)
-    if (retrievedPayment instanceof Error) return retrievedPayment
-    expect(retrievedPayment.status).toBe(PaymentStatus.Settled)
-    if (retrievedPayment.status !== PaymentStatus.Settled) return
-    expect(retrievedPayment.confirmedDetails?.revealedPreImage).toBe(revealedPreImage)
-
-    // Run delete-payments cronjob
-    const timestamp2Months = new Date(Date.now() - TWO_MONTHS_IN_MS)
-    expect(Number(timestamp2Months)).toBeLessThan(Number(retrievedPayment.createdAt))
-    const deleteLnPayments1Hour = await Lightning.deleteLnPaymentsBefore(timestamp2Months)
-    if (deleteLnPayments1Hour instanceof Error) throw deleteLnPayments1Hour
-
-    // Confirm payment still exists
-    retrievedPayment = await lndService.lookupPayment({ paymentHash })
-    expect(retrievedPayment).not.toBeInstanceOf(Error)
-    if (retrievedPayment instanceof Error) return retrievedPayment
-    expect(retrievedPayment.status).toBe(PaymentStatus.Settled)
-    if (retrievedPayment.status !== PaymentStatus.Settled) return
-    expect(retrievedPayment.confirmedDetails?.revealedPreImage).toBe(revealedPreImage)
-
-    // Run updateLnPayments task
-    const lnPaymentUpdateOnPending = await Lightning.updateLnPayments()
-    if (lnPaymentUpdateOnPending instanceof Error) throw lnPaymentUpdateOnPending
-
-    // Run delete-payments cronjob again for payments before 2 weeks ago
-    const deleteLnPayments1HourRetry = await Lightning.deleteLnPaymentsBefore(
-      timestamp2Months,
-    )
-    if (deleteLnPayments1HourRetry instanceof Error) throw deleteLnPayments1HourRetry
-
-    // Confirm payment still exists
-    retrievedPayment = await lndService.lookupPayment({ paymentHash })
-    expect(retrievedPayment).not.toBeInstanceOf(Error)
-    if (retrievedPayment instanceof Error) return retrievedPayment
-    expect(retrievedPayment.status).toBe(PaymentStatus.Settled)
-    if (retrievedPayment.status !== PaymentStatus.Settled) return
-    expect(retrievedPayment.confirmedDetails?.revealedPreImage).toBe(revealedPreImage)
-
-    // Run delete-payments cronjob again for all payments
-    const timestampNow = new Date(Date.now())
-    expect(Number(timestampNow)).toBeGreaterThan(Number(retrievedPayment.createdAt))
-    const deleteLnPayments = await Lightning.deleteLnPaymentsBefore(timestampNow)
-    if (deleteLnPayments instanceof Error) throw deleteLnPayments
-
-    // Confirm payment was deleted
-    const retrievedDeletedPayment = await lndService.lookupPayment({ paymentHash })
-    expect(retrievedDeletedPayment).toBeInstanceOf(PaymentNotFoundError)
   })
 })
